+++ resources/lang/en/blog.php
@@ -3,7 +3,7 @@
 return [
 
     'blog_empty' => 'Create there your adventures!',
-    'new_blog' => 'New adventure',
+    'new_blog_g' => 'New group adventure',
     'adventure_name' => 'Name of the adventure',
     'empty_posts' => 'It seems that you haven\'t writen anything yet. Are you in the mood now?',
     'write_post' => 'Write a post',